user story #561048 : TFS - Log improvements
