+++ OctaneManager/Properties/AssemblyInfo.cs
@@ -47,5 +47,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.213")]
-[assembly: AssemblyFileVersion("1.0.0.213")]
+[assembly: AssemblyVersion("1.0.0.223")]
+[assembly: AssemblyFileVersion("1.0.0.223")]
