+++ OctaneManager/RestServer/Views/config.html
@@ -7,6 +7,7 @@
 	<meta name="author" content="Microfocus">
 
 	<link rel="stylesheet" type="text/css" href="resources/style.css">
+	<link rel="shortcut icon" href="resources/f-ci-tfs.png">
 	<style>
 
 
@@ -95,15 +96,8 @@ table .itemCaption{
 
 	<div id="header">
 		<img id="octaneIcon" src="resources/s-octane-logo.svg" alt="ALM Octane">
+		<a href="../" id="home"><img class="linkIcon" src="resources/home-white.svg" alt="Home"></a>
 		<div id="title">TFS Plugin for ALM Octane - Configuration</div>
-	</div>
-
-	
-	<div id="links">
-		<h2>Links</h2>
-		<a href="../"><img class="linkIcon" src="resources/home.svg" alt="Home"><span class="linkText">Home</span></a>
-
-		
 	</div>
