+++ OctaneManager/RestServer/Views/console.html
@@ -1,4 +1,5 @@
 ﻿<!doctype html>
+<!doctype html>
 <html lang="en">
 <head>
 	<meta charset="utf-8">
@@ -7,6 +8,7 @@
 	<meta name="author" content="Microfocus">
 
 	<link rel="stylesheet" type="text/css" href="resources/style.css">
+	<link rel="shortcut icon" href="resources/f-ci-tfs.png">
 	<style>
 
 
@@ -56,7 +58,7 @@ table{
 		<h2>Links</h2>
 
 		<a href="/config"><img class="linkIcon" src="resources/configuration.svg" alt="Configuration"><span class="linkText">Configuration</span></a>
-		<a href="/logs/app/last" target="_blank"><img class="linkIcon" src="resources/f-step-simple.svg" alt="Application Log" ><span class="linkText">Application Log</span></a>
+		<a href="/logs"><img class="linkIcon" src="resources/f-step-simple.svg" alt="Logs" ><span class="linkText">Logs</span></a>
 
 		<div id="actions" class="hidden">
 			<h2>Actions</h2>
